Auto stash before merge of "main" and "origin/main"

+++ Assets/Highest.cs
@@ -12,38 +12,38 @@ public class Highest : MonoBehaviour
     void Start()
     {
 
-        paso  = PlayerPrefs.GetInt("HighScore", 0).ToString();
+        paso  = PlayerPrefs.GetFloat("HighScore", 0).ToString();
         highScore.text = paso;
     }
 
     public void ReviewData()
     {
-        int DataToReview = PlayerPrefs.GetInt("Temporary Points");
-        int tmp1, tmp2, tmp3;
+        float DataToReview = PlayerPrefs.GetFloat("Temporary Points");
+        float tmp1, tmp2, tmp3;
 
-        if (DataToReview > PlayerPrefs.GetInt("HighScore"))
+        if (DataToReview > PlayerPrefs.GetFloat("HighScore"))
         {
-            tmp1 = PlayerPrefs.GetInt("HighScore");
-            PlayerPrefs.SetInt("HighScore", DataToReview);
+            tmp1 = PlayerPrefs.GetFloat("HighScore");
+            PlayerPrefs.SetFloat("HighScore", DataToReview);
         }
         else
         {
             tmp1 = DataToReview;
         }
 
-        if (tmp1 > PlayerPrefs.GetInt("SecondHighest"))
+        if (tmp1 > PlayerPrefs.GetFloat("SecondHighest"))
         {
-            tmp2 = PlayerPrefs.GetInt("SecondHighest");
-            PlayerPrefs.SetInt("SecondHighest", tmp1);
+            tmp2 = PlayerPrefs.GetFloat("SecondHighest");
+            PlayerPrefs.SetFloat("SecondHighest", tmp1);
         }
         else
         {
             tmp2 = tmp1;
         }
-        if (tmp2 > PlayerPrefs.GetInt("ThirdHighest"))
+        if (tmp2 > PlayerPrefs.GetFloat("ThirdHighest"))
         {
-            tmp3 = PlayerPrefs.GetInt("ThirdHighest");
-            PlayerPrefs.SetInt("ThirdHighest",tmp2);
+            tmp3 = PlayerPrefs.GetFloat("ThirdHighest");
+            PlayerPrefs.SetFloat("ThirdHighest",tmp2);
         }
         else
         {
